[iCub] rename test
